hotfix: changing WSDL url when is over SSL
The SOAP library is defined by default to make requests in the HTTP protocol. When an SSL URL is given, the PHP core changes the protocol internally to HTTP, breaking the communication.
a
This change fix that, in the construction of the client, we check if the WSDL is over the SSL protocol, if so, in the `__doRequest` method, we change the protocol from HTTP to HTTPS, so the cURL can be executed correctly.